Set outputted class files from JavaCompilationBox in form of wildcards.
Java, Java, Java... Java is very special in its compilation pipeline and
it is not predictable at all which files are outputted during
compilation. Because it might output even separate classfiles for
each nested class, which is hard to predict. This is causing us a lot of trouble, which could be avoided if Java compilation would behave in a deterministic way.